Move task detail logic into popup component
